nekjfajkdna

+++ resources/views/livewire/send-sms.blade.php
@@ -52,14 +52,20 @@
     <div class="phone-varification">
       <div class="phone-inputBox">
         <label for="phone">Telefon *</label>
-        <input type="number" value="{{old('phone]')}}" <?=($errors->has('phone')) ? 'style="border: 1px solid rgb(245, 70, 70)"' : ''?> name="phone" wire:model.lazy="phone" id="phone" onfocusout="myFunction()" oninput="javascript: if (this.value.length > this.maxLength) this.value = this.value.slice(0, this.maxLength);" placeholder="+998 xx-xxx-xx-xx" maxlength="9"
-        autocomplete="off">
-        <div id="countdownDisplay">01:00</div>
+        <input type="tel" value="{{old('phone]')}}" <?=($errors->has('phone')) ? 'style="border: 1px solid rgb(245, 70, 70)"' : ''?> name="phone" wire:model.lazy="phone" oninput="javascript: if (this.value.length > this.maxLength) this.value = this.value.slice(0, this.maxLength);" 
+        maxlength="9" id="phone" onfocusout="myFunction(this)" placeholder="+998 xx-xxx-xx-xx" autocomplete="off" 
+        >
+        {{-- oninput="javascript: if (this.value.length > this.maxLength) this.value = this.value.slice(0, this.maxLength);" 
+        maxlength="9"
+        --}}
+
       </div>
+      
       <div class="varification-code_inputBox">
         <label for="verifyCode">Tasdiqlash kodi *</label>
         <input type="number" name="verifyCode" oninput="javascript: if (this.value.length > this.maxLength) this.value = this.value.slice(0, this.maxLength);" maxlength="4" id="verifyCode" />
       </div>
+      <p class="timer" id="countdownDisplay">01:00</p>
       
     {{-- <button id="startButton">Start Countdown</button> --}}
 
@@ -72,28 +78,68 @@
 </div>
 
 
+{{-- <script src="{{asset('frontend/js/components/form.js')}}" type="module"></script> --}}
+<script src="https://cdn.jsdelivr.net/npm/imask"></script>
+<script type="module">
+  const button = document.getElementById("file-upload_btn");
+  const fileUpload = document.getElementById("file");
+
+  button.addEventListener("click", (e) => {
+    fileUpload.click();
+    e.preventDefault();
+  });
+
+  fileUpload.addEventListener("click", (event) => {
+    event.stopPropagation();
+  });
+
+  // Varification code input
+  const varifyCodeInput = document.querySelector("#verifyCode");
+  varifyCodeInput.addEventListener(
+    "keypress",
+    (evt) => ["e", "E", "+", "-"].includes(evt.key) && evt.preventDefault()
+  );
+
+</script>
 <script>
+  
+  // const phoneInput = document.getElementById("phone");
+  // function func(e){
+  //     e.value = "998";
+  //     setTimeout(() => {
+  //       phoneInput.setSelectionRange(6, 6);
+  //     }, 10);
+
+  //     IMask(phoneInput, {
+  //       mask: "998000000000",
+  //       lazy: false,
+  //     });
+  // }
+
   const countdownDisplay = document.getElementById('countdownDisplay');
   let countdownInterval;
 
-    function myFunction() {
-      clearInterval(countdownInterval);
-      
-      // Set the countdown time
-      let timeLeft = 59;
-      countdownDisplay.textContent = '00:'+timeLeft;
-      
-      // Start the countdown
-      countdownInterval = setInterval(function() {
-          timeLeft--;
+    function myFunction(e) {
+      console.log(typeof  e.value);
+      if(e.value.length==9){
+          clearInterval(countdownInterval);
+          
+          // Set the countdown time
+          let timeLeft = 59;
           countdownDisplay.textContent = '00:'+timeLeft;
           
-          if (timeLeft <= 0) {
-            sessionStorage.removeItem('verification_code'); 
-            sessionStorage.clear();  
-            clearInterval(countdownInterval);
-          }
-      }, 1000);
+          // Start the countdown
+          countdownInterval = setInterval(function() {
+              timeLeft--;
+              countdownDisplay.textContent = '00:'+timeLeft;
+              
+              if (timeLeft <= 0) {
+                sessionStorage.removeItem('verification_code'); 
+                sessionStorage.clear();  
+                clearInterval(countdownInterval);
+              }
+          }, 1000);
+      }
     }
 
 </script>
